+++ client/src/app/home/ui/tweet/tweet-item/tweet-item.component.ts
@@ -1,4 +1,13 @@
-import { Component, Input, OnInit, ViewEncapsulation } from '@angular/core';
+import {
+  Component,
+  EventEmitter,
+  Input,
+  OnChanges,
+  OnInit,
+  Output,
+  SimpleChanges,
+  ViewEncapsulation,
+} from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { RetweetService } from 'src/app/home/services/retweet.service';
@@ -15,11 +24,19 @@ import { PostMakerDialogComponent } from '../post-maker-dialog/post-maker-dialog
   styleUrls: ['./tweet-item.component.css'],
   encapsulation: ViewEncapsulation.None,
 })
-export class TweetItemComponent implements OnInit {
+export class TweetItemComponent implements OnInit, OnChanges {
   @Input() data?: TweetModel;
+  @Input() detailedView: boolean = false;
+
+  @Output() onEditClick = new EventEmitter<TweetModel>();
+  @Output() onDeleteClick = new EventEmitter<TweetModel>();
+  @Output() onLikeClick = new EventEmitter<TweetModel>();
+  @Output() onCommentClick = new EventEmitter<TweetModel>();
+  @Output() onRetweetClick = new EventEmitter<TweetModel>();
+  @Output() onQuoteRetweetClick = new EventEmitter<TweetModel>();
+  @Output() onRetweetUndoClick = new EventEmitter<TweetModel>();
 
   tweet?: TweetModel;
-
   userAuth?: UserModel;
 
   alreadyLiked: boolean = false;
@@ -33,6 +50,11 @@ export class TweetItemComponent implements OnInit {
     private router: Router,
     public dialog: MatDialog
   ) {}
+
+  ngOnChanges(changes: SimpleChanges): void {
+    if (this.data?.isRetweet) this.tweet = this.data?.originalTweet;
+    else this.tweet = this.data;
+  }
 
   ngOnInit(): void {
     this.userService.userAuth.subscribe((res) => {
@@ -52,7 +74,8 @@ export class TweetItemComponent implements OnInit {
   }
 
   onTweetClick() {
-    this.router.navigate([`/home/tweet/${this.tweet?.id}`]);
+    if (!this.detailedView)
+      this.router.navigate([`/home/tweet/${this.tweet?.id}`]);
   }
 
   onTagClick(hashtag: string) {
@@ -62,32 +85,34 @@ export class TweetItemComponent implements OnInit {
   }
 
   onLike() {
-    this.tweetService
-      .likeTweet(this.tweet?.id!, !this.alreadyLiked)
-      .subscribe();
-
-    if (this.alreadyLiked && this.tweet) {
+    if (this.alreadyLiked && this.tweet)
       this.tweet.likes = this.tweet.likes.filter(
         (likedBy) => likedBy !== this.userAuth?.userId
       );
-    } else {
-      this.tweet?.likes?.push(this.userAuth?.userId!);
-    }
+    else this.tweet?.likes?.push(this.userAuth?.userId!);
 
     this.alreadyLiked = !this.alreadyLiked;
+
+    this.tweetService.likeTweet(this.tweet?.id!, this.alreadyLiked).subscribe();
   }
 
   onComment() {
     this.onTweetClick();
+    this.onCommentClick.emit(this.tweet);
   }
 
   onRetweet() {
+    this.tweet?.retweetUsers?.push(this.userAuth?.userId!);
+    this.alreadyRetweeted = true;
+
     this.retweetService
       .createRetweet({
         isQuoteRetweet: false,
         originalTweetId: this.tweet?.id!,
       })
       .subscribe();
+
+    this.onRetweetClick.emit(this.tweet);
   }
 
   onQuote() {
@@ -108,18 +133,25 @@ export class TweetItemComponent implements OnInit {
             text: result.text,
             hashtags: result.hashtags!,
           })
-          .subscribe();
+          .subscribe((res) => this.tweet?.quoteRetweets?.push(res.id));
       }
     });
+
+    this.onQuoteRetweetClick.emit(this.tweet);
   }
 
   onRetweetUndo() {
+    this.tweet!.retweetUsers = this.tweet?.retweetUsers?.filter(
+      (retweetBy) => retweetBy !== this.userAuth?.userId
+    )!;
+    this.alreadyRetweeted = false;
+
     this.retweetService.undoRetweet(this.tweet?.id!).subscribe(() => {
-      if (this.data?.isRetweet) {
-        this.storeService.removeTweetFromTweetList(this.data?.id!);
-      } else {
-        this.storeService.tweetList.subscribe((res) => {
-          res?.forEach((tweet) => {
+      if (!this.detailedView) {
+        if (this.data?.isRetweet) {
+          this.storeService.removeTweetFromTweetList(this.data?.id!);
+        } else {
+          this.storeService.tweetList.getValue().forEach((tweet) => {
             if (
               tweet.isRetweet &&
               tweet.originalTweetId === this.tweet?.id &&
@@ -128,9 +160,11 @@ export class TweetItemComponent implements OnInit {
               this.storeService.removeTweetFromTweetList(tweet.id);
             }
           });
-        });
+        }
       }
     });
+
+    this.onRetweetUndoClick.emit(this.data);
   }
 
   onEdit() {
@@ -153,9 +187,13 @@ export class TweetItemComponent implements OnInit {
             text: result.text,
             hashtags: result.hashtags,
           })
-          .subscribe();
+          .subscribe((res) => {
+            this.tweet = res;
+          });
       }
     });
+
+    this.onEditClick.emit(this.tweet);
   }
 
   onDelete() {
@@ -166,5 +204,7 @@ export class TweetItemComponent implements OnInit {
           .subscribe();
       else this.tweetService.deleteTweet(this.tweet?.id).subscribe();
     }
+
+    this.onDeleteClick.emit(this.tweet);
   }
 }
